fix: updated empty files and added rootstack

+++ client/src/Store/actions/users.ts
@@ -18,38 +18,50 @@ import {
 import { Book } from '../../Interfaces/bookObject';
 import { User } from '../../Interfaces/userObject';
 
-const accessToken: string = localStorage.getItem('accessToken');
+const accessToken: string | null = localStorage.getItem('accessToken');
 
 export const _loadDashboard = () => async (dispatch: AppDispatch) => {
-  const userDashboard = await loadDashboard(accessToken);
-  dispatch({ type: LOAD_DASHBOARD, payload: userDashboard });
+  if (accessToken) {
+    const userDashboard = await loadDashboard(accessToken);
+    dispatch({ type: LOAD_DASHBOARD, payload: userDashboard });
+  }
 };
 
 export const _getUserWithBooks = () => async (dispatch: AppDispatch) => {
-  const { userWithBooks } = await getUserWithBooks(accessToken);
-  dispatch({ type: GET_USER_WITH_BOOKS, payload: userWithBooks });
+  if (accessToken) {
+    const { userWithBooks } = await getUserWithBooks(accessToken);
+    dispatch({ type: GET_USER_WITH_BOOKS, payload: userWithBooks });
+  }
 };
 
 export const _addSavedBook = (book: Book) => async (dispatch: AppDispatch) => {
-  const { savedBook } = await addSavedBook(accessToken, book);
-  dispatch({ type: ADD_SAVED_BOOK, payload: savedBook });
+  if (accessToken) {
+    const { savedBook } = await addSavedBook(accessToken, book);
+    dispatch({ type: ADD_SAVED_BOOK, payload: savedBook });
+  }
 };
 
 export const _deleteSavedBook = (book: Book) => async (
   dispatch: AppDispatch
 ) => {
-  await deleteSavedBook(accessToken, book);
-  dispatch({ type: DELETE_SAVED_BOOK, payload: book });
+  if (accessToken) {
+    await deleteSavedBook(accessToken, book);
+    dispatch({ type: DELETE_SAVED_BOOK, payload: book });
+  }
 };
 
 export const _updateRating = (book: Book, rating: number) => async (
   dispatch: AppDispatch
 ) => {
-  const { savedBook } = await updateRating(accessToken, book, rating);
-  dispatch({ type: UPDATE_RATING, payload: savedBook });
+  if (accessToken) {
+    const { savedBook } = await updateRating(accessToken, book, rating);
+    dispatch({ type: UPDATE_RATING, payload: savedBook });
+  }
 };
 
 export const _deleteRating = (book: Book) => async (dispatch: AppDispatch) => {
-  await deleteRating(accessToken, book);
-  dispatch({ type: DELETE_RATING, payload: book });
+  if (accessToken) {
+    await deleteRating(accessToken, book);
+    dispatch({ type: DELETE_RATING, payload: book });
+  }
 };
